Move createMTOServiceItem templates into directory
