+++ src/constants/MoveHistory/EventTemplates/CreateMTOServiceItem/createMTOServiceItem.test.js
@@ -2,26 +2,33 @@ import getTemplate from 'constants/MoveHistory/TemplateManager';
 import o from 'constants/MoveHistory/UIDisplay/Operations';
 import a from 'constants/MoveHistory/Database/Actions';
 import t from 'constants/MoveHistory/Database/Tables';
-import e from 'constants/MoveHistory/EventTemplates/createMTOServiceItemCustomerContacts';
+import e from 'constants/MoveHistory/EventTemplates/CreateMTOServiceItem/createMTOServiceItem';
 
-describe('when given a Create basic service item customer contacts history record', () => {
+describe('when given a Create basic service item history record', () => {
   const item = {
     action: a.INSERT,
     changedValues: {
-      first_available_delivery_date: '2022-06-30T00:00:00+00:00',
-      time_military: '1500Z',
-      type: 'SECOND',
+      reason: 'Test',
+      status: 'SUBMITTED',
     },
+    context: [
+      {
+        name: 'Domestic uncrating',
+        shipment_type: 'HHG',
+      },
+    ],
     eventName: o.createMTOServiceItem,
-    tableName: t.mto_service_item_customer_contacts,
+    tableName: t.mto_service_items,
   };
-  it('correctly matches the create service item customer contacts event', () => {
+  it('correctly matches the create service item event', () => {
     const result = getTemplate(item);
     expect(result).toMatchObject(e);
     expect(result.getEventNameDisplay()).toEqual('Requested service item');
     expect(result.getDetailsLabeledDetails(item)).toMatchObject({
-      first_available_delivery_date: '2022-06-30T00:00:00+00:00',
-      second_available_delivery_time: '1500Z',
+      service_item_name: 'Domestic uncrating',
+      shipment_type: 'HHG',
+      reason: 'Test',
+      status: 'SUBMITTED',
     });
   });
 });
