+++ src/constants/MoveHistory/EventTemplates/CreateMTOServiceItem/createMTOServiceItemUpdateMoveStatus.test.js
@@ -1,7 +1,7 @@
 import getTemplate from 'constants/MoveHistory/TemplateManager';
 import d from 'constants/MoveHistory/UIDisplay/DetailsTypes';
 import o from 'constants/MoveHistory/UIDisplay/Operations';
-import e from 'constants/MoveHistory/EventTemplates/createMTOServiceItemUpdateMoveStatus';
+import e from 'constants/MoveHistory/EventTemplates/CreateMTOServiceItem/createMTOServiceItemUpdateMoveStatus';
 
 describe('when given a move status update with create mto service item history record', () => {
   const item = {
