+++ src/constants/MoveHistory/EventTemplates/index.js
@@ -4,9 +4,9 @@ export { default as approveShipmentDiversion } from './approveShipmentDiversion'
 export { default as createBasicServiceItem } from './createBasicServiceItem';
 export { default as createEntitlements } from './CreateOrders/createEntitlements';
 export { default as createMoves } from './CreateOrders/createMoves';
-export { default as createMTOServiceItem } from './createMTOServiceItem';
-export { default as createMTOServiceItemCustomerContacts } from './createMTOServiceItemCustomerContacts';
-export { default as createMTOServiceItemDimensions } from './createMTOServiceItemDimensions';
+export { default as createMTOServiceItem } from './CreateMTOServiceItem/createMTOServiceItem';
+export { default as createMTOServiceItemCustomerContacts } from './CreateMTOServiceItem/createMTOServiceItemCustomerContacts';
+export { default as createMTOServiceItemDimensions } from './CreateMTOServiceItem/createMTOServiceItemDimensions';
 export { default as createMTOShipment } from './createMTOShipment';
 export { default as createMTOShipmentAddresses } from './createMTOShipmentAddresses';
 export { default as createMTOShipmentAgent } from './createMTOShipmentAgent';
@@ -35,7 +35,7 @@ export { default as updateMTOShipmentAddresses } from './updateMTOShipmentAddres
 export { default as updateMTOShipmentAgent } from './updateMTOShipmentAgent';
 export { default as updateMTOShipmentDeprecatePaymentRequest } from './updateMTOShipmentDeprecatePaymentRequest';
 export { default as reweighPaymentRequest } from './reweighPaymentRequest';
-export { default as updateMTOServiceItemAddress } from './updateMTOServiceItemAddress';
+export { default as updateMTOServiceItemAddress } from './CreateMTOServiceItem/updateMTOServiceItemAddress';
 export { default as updateOrder } from './updateOrder';
 export { default as updatePaymentRequest } from './updatePaymentRequest';
 export { default as updatePaymentRequestJobRunner } from './updatePaymentRequestJobRunner';
@@ -45,7 +45,7 @@ export { default as updateServiceItemStatus } from './updateServiceItemStatus';
 export { default as updateMTOReviewedBillableWeightsAt } from './updateMTOReviewedBillableWeightsAt';
 export { default as uploadAmendedOrders } from './uploadAmendedOrders';
 export { default as updateMTOShipmentStatus } from './updateMTOShipmentStatus';
-export { default as updateMTOServiceItemMoveStatus } from './createMTOServiceItemUpdateMoveStatus';
+export { default as updateMTOServiceItemMoveStatus } from './CreateMTOServiceItem/createMTOServiceItemUpdateMoveStatus';
 export { default as updateMoveEstimatedExcessWeight } from './updateMTOShipmentPrimeEstimatedExcessWeight';
 export { default as updateMTOStatusServiceCounselingCompleted } from './updateMTOStatusServiceCounselingCompleted';
 export { default as updateAllowanceCounseling } from './updateAllowanceCounseling';
